Conflicts:
	src/main/java/com/comdosoft/financial/manage/domain/zhangfu/Merchant.java
	src/main/resources/mapper/zhangfu/CustomerMapper.xml
	src/main/resources/mapper/zhangfu/TerminalMapper.xml

+++ src/main/java/com/comdosoft/financial/manage/mapper/zhangfu/GoodMapper.java
@@ -44,9 +44,10 @@ public interface GoodMapper {
 			
 	long countByKeys(@Param("status") Byte status, @Param("keys") String keys);
 	
-	Good findGoodInfo(Long id);
+	Good findGoodInfo(Integer id);
 	
-	Good findPageRowGood(Long id);
+	Good findPageRowGood(Integer id);
 
 	List<Good> selectByStatusAndName(@Param("status") Byte status, @Param("keys") String keys);
+	
 }
